Workarounds
BUILD: Bump version
FEAT: Added new HUD workaround for multiplayer
REFACTOR: Updated fade in workaround's description

So it turns out that the HUD and fading in text only breaks in a very specific environment. I was using a server software running on 1.17, so that might be the cause right there. Testing the same software running on 1.19 worked perfectly fine

Signed-off-by: Superkat32 <superkat1806@gmail.com>